chore: change a middleware from one archive to the right one
